Move integration serde serializer tests to the integration tests directory
